+++ src/redmine/models/journal.ts
@@ -1,0 +1,8 @@
+import { NamedEntity } from "./named-entity";
+
+export interface Journal {
+  id: number;
+  user: NamedEntity;
+  notes: string;
+  created_on: string;
+}