+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "vscode-redmine" extension will be documented in this
 
 ## [Unreleased]
 
+### Added
+
+- [Issue #49](https://github.com/rozpuszczalny/vscode-redmine/issues/49) - filter and sort properties were added into config
+- Grabbing description and journals of issue feature was added
+
 ## 1.1.1 - 17.08.2022
 
 ### Fixed